refactor: move back methods & constants to config_editor
