+++ tui/src/ui/model/update.rs
@@ -2,20 +2,16 @@ use std::path::PathBuf;
 use std::time::Duration;
 
 use anyhow::{Result, anyhow};
-use termusiclib::config::v2::tui::theme::ThemeColors;
 use termusiclib::player::{
     PlayerProgress, RunningStatus, UpdateEvents, UpdatePlaylistEvents, clamp_u16,
 };
 use termusiclib::podcast::{PodcastDLResult, PodcastSyncResult};
 use termusiclib::track::MediaTypesSimple;
-use termusiclib::utils::get_app_config_path;
 use tokio::runtime::Handle;
 use tokio::time::sleep;
-use tuirealm::event::Event;
 use tuirealm::props::{AttrValueRef, Attribute, QueryResult};
 
 use crate::ui::ids::Id;
-use crate::ui::model::UserEvent;
 use crate::ui::model::youtube_options::YTDLMsg;
 use crate::ui::msg::{
     CoverDLResult, DBMsg, DeleteConfirmMsg, ErrorPopupMsg, GSMsg, HelpPopupMsg, LIMsg, LyricMsg,
@@ -1271,52 +1267,5 @@ impl Model {
         }
 
         Ok(())
-    }
-
-    /// Re-apply theme colors to components that bake style in at mount time
-    /// and don't otherwise re-read `config_tui` on every render (unlike e.g. Playlist).
-    pub fn refresh_static_theme_components(&mut self) {
-        let ev = Event::User(UserEvent::Forward(Msg::ReloadTheme));
-
-        for id in [Id::Library, Id::Podcast, Id::Episode, Id::MessagePopup] {
-            if let Some(component) = self.app.get_component_mut(&id) {
-                component.on(&ev);
-            }
-        }
-    }
-
-    pub(crate) fn resolve_theme_by_index(&self, index: usize) -> Option<ThemeColors> {
-        if index == 0 {
-            return Some(ThemeColors::full_default());
-        }
-        if index == 1 {
-            return Some(ThemeColors::full_native());
-        }
-
-        let theme_filename = self
-            .config_editor
-            .themes
-            .get(index - ThemeColors::THEMES_WITHOUT_FILES)?;
-
-        match get_app_config_path() {
-            Ok(mut theme_path) => {
-                theme_path.push("themes");
-                theme_path.push(format!("{theme_filename}.yml"));
-                match ThemeColors::from_yaml_file(&theme_path) {
-                    Ok(mut theme) => {
-                        theme.file_name = Some(theme_filename.clone());
-                        Some(theme)
-                    }
-                    Err(e) => {
-                        error!("Failed to load theme colors: {e:?}");
-                        None
-                    }
-                }
-            }
-            Err(e) => {
-                error!("Error getting config path: {e:?}");
-                None
-            }
-        }
     }
 }
